scripts/requirements.txt
openai>=1.12.0
supabase>=2.3.0
pdfminer.six>=20221105
python-dotenv>=1.0.0
tqdm>=4.65.0

